+++ flowmachine/flowmachine/core/server/query_schemas/flowmachine_query.py
@@ -40,6 +40,8 @@ from .unmoving_at_reference_location_counts import (
     UnmovingAtReferenceLocationCountsSchema,
 )
 from .unmoving_counts import UnmovingCountsSchema
+from .labelled_spatial_aggregate import LabelledSpatialAggregateSchema
+from .labelled_flows import LabelledFlowsSchema
 
 
 class FlowmachineQuerySchema(OneOfSchema):
@@ -66,6 +68,8 @@ class FlowmachineQuerySchema(OneOfSchema):
         "unmoving_counts": UnmovingCountsSchema,
         "unmoving_at_reference_location_counts": UnmovingAtReferenceLocationCountsSchema,
         "trips_od_matrix": TripsODMatrixSchema,
+        "labelled_spatial_aggregate": LabelledSpatialAggregateSchema,
+        "labelled_flows": LabelledFlowsSchema,
     }