+++ integration_tests/tests/flowmachine_server_tests/test_server.py
@@ -78,6 +78,8 @@ def test_get_available_queries(zmq_host, zmq_port):
                 "unmoving_counts",
                 "unmoving_at_reference_location_counts",
                 "trips_od_matrix",
+                "labelled_spatial_aggregate",
+                "labelled_flows",
             ]
         },
     }